requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# 3D geometry and calculations
shapely>=2.0.0
trimesh>=3.20.0
pyproj>=3.6.0
open3d>=0.18.0  # Advanced 3D processing for better window detection
pygltflib>=1.15.0  # GLB/glTF scene graph parsing
pyglet<2  # Required for trimesh viewer fallback (trimesh requires pyglet<2)

# GUI (choose one or both)
PyQt6>=6.5.0
# or
# tkinter (built-in with Python)

# 3D Viewer (optional - for GLB 3D viewer window)
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0

# Report generation
matplotlib>=3.7.0
reportlab>=4.0.0
pillow>=10.0.0

# BIM/IFC support
ifcopenshell>=0.7.0

# Date/time calculations for sun position
astral>=3.2
pytz>=2023.3

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

